+++ src/ffi_c.cpp
@@ -274,6 +274,9 @@ GDK_DEFINE_C_FUNCTION_3(GA_get_assets, struct GA_session*, session, const GA_jso
 GDK_DEFINE_C_FUNCTION_3(GA_validate_asset_domain_name, struct GA_session*, session, const GA_json*, params, GA_json**,
     output, { *json_cast(output) = new nlohmann::json(session->validate_asset_domain_name(*json_cast(params))); });
 
+GDK_DEFINE_C_FUNCTION_3(GA_validate, struct GA_session*, session, const GA_json*, details, struct GA_auth_handler**,
+    call, { *call = make_call(new ga::sdk::validate_call(*session, *json_cast(details))); })
+
 GDK_DEFINE_C_FUNCTION_4(GA_register_user, struct GA_session*, session, const GA_json*, hw_device, const GA_json*,
     details, struct GA_auth_handler**, call,
     { *call = make_call(new ga::sdk::register_call(*session, *json_cast(hw_device), *json_cast(details))); })