+++ src/swap_auth_handlers.cpp
@@ -295,5 +295,33 @@ namespace sdk {
             GDK_RUNTIME_ASSERT_MSG(false, "Unknown next handler called");
         }
     }
+
+    //
+    // Validate
+    //
+    validate_call::validate_call(session& session, const nlohmann::json& details)
+        : auth_handler_impl(session, "validate")
+        , m_details(details)
+    {
+    }
+
+    auth_handler::state_type validate_call::call_impl()
+    {
+        m_result["errors"] = nlohmann::json::array();
+        m_result["is_valid"] = false;
+        try {
+            liquidex_impl();
+            m_result["is_valid"] = true;
+        } catch (const std::exception& e) {
+            m_result["errors"].emplace_back(e.what());
+        }
+        return state_type::done;
+    }
+
+    void validate_call::liquidex_impl()
+    {
+        const auto& proposal = m_details.at("liquidex_v0").at("proposal");
+        liquidex_validate_proposal(proposal);
+    }
 } // namespace sdk
 } // namespace ga